requirements.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.12.2
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==2.1.1
cloudinary==1.25.0
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==22.1
django-rest-auth==0.9.5
django-storages==1.13.1
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.2.0
env==0.1.0
filelock==3.12.3
flake8==5.0.4
geographiclib==2.0
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
idna==3.3
mccabe==0.7.0
msgpack==1.0.7
oauthlib==3.2.1
Pillow==9.2.0
platformdirs==3.10.0
protobuf==4.21.5
psycopg2==2.9.6
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
redis==5.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.2
toml==0.10.2
types-cryptography==3.3.23
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.24.4
whitenoise==6.4.0
